backend/requirements.txt
fastapi
uvicorn
langchain
openai
pydantic
python-dotenv
sqlalchemy
psycopg2-binary
langchain-openai
